+++ src/Mongo/test/Eventuous.Tests.Projections.MongoDB/Fixtures/DomainFixture.cs
@@ -6,14 +6,15 @@ using static Eventuous.Tests.Projections.MongoDB.Fixtures.IntegrationFixture;
 namespace Eventuous.Tests.Projections.MongoDB.Fixtures;
 
 public static class DomainFixture {
-    static DomainFixture() => TypeMap.RegisterKnownEventTypes();
+    static DomainFixture()
+        => TypeMap.RegisterKnownEventTypes();
 
     public static BookingImported CreateImportBooking() {
         var from = Instance.Auto.Create<DateTime>();
 
         return new BookingImported(
             Instance.Auto.Create<string>(),
-            Instance.Auto.Create<decimal>(),
+            Instance.Auto.Create<float>(),
             LocalDate.FromDateTime(from),
             LocalDate.FromDateTime(from.AddDays(Instance.Auto.Create<int>()))
         );
@@ -25,8 +26,8 @@ public record BookingDocument(string Id) : ProjectedDocument(Id) {
     public string    RoomId       { get; init; } = null!;
     public LocalDate CheckInDate  { get; init; }
     public LocalDate CheckOutDate { get; init; }
-    public decimal   BookingPrice { get; init; }
-    public decimal   PaidAmount   { get; init; }
-    public decimal   Outstanding  { get; init; }
+    public float     BookingPrice { get; init; }
+    public float     PaidAmount   { get; init; }
+    public float     Outstanding  { get; init; }
     public bool      Paid         { get; init; }
 }
